backend/requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# ML - pre-built wheels
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
joblib==1.3.2
xgboost==2.0.3

# API & Services
google-generativeai==0.3.2
pydantic[email]==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
email-validator==2.1.0

# Utils
python-dotenv==1.0.0
aiofiles==23.2.1
